feat: Sanitize service ID in tool name generation
The service ID was not being sanitized before being used to generate the tool name. This could lead to invalid tool names if the service ID contained invalid characters.

This commit fixes the issue by sanitizing the service ID before it is used to generate the tool name. A new test case has been added to verify the fix.