feat: Add end-to-end test for ipinfo.io service
This commit introduces a new end-to-end test for the ipinfo.io popular service.

The test verifies that the `ipinfo.io` tool can be called and that the response contains the expected output. It also ensures that the `IPINFO_API_TOKEN` is correctly used for authentication.

In addition, this commit fixes a bug in the `ipinfo.io` service configuration where the authentication details were placed incorrectly within the `http_service` block instead of at the top level as `upstream_authentication`.